Update reading csv

+++ src/model/Player.java
@@ -5,15 +5,15 @@ public class Player {
 	private String nombre;
 	private int edad;
 	private String equipo;
-	private int puntosPorPartido;
-	private int rebotesPorPartido;
-	private int asistenciasPorPartido;
-	private int robosPorPartido;
-	private int bloqueosPorPartido;
-	private int faltasPorPartido;
+	private double puntosPorPartido;
+	private double rebotesPorPartido;
+	private double asistenciasPorPartido;
+	private double robosPorPartido;
+	private double bloqueosPorPartido;
+	private double faltasPorPartido;
 	
-	public Player(String nombre, int edad, String equipo, int puntosPorPartido, int rebotesPorPartido,
-			int asistenciasPorPartido, int robosPorPartido, int bloqueosPorPartido, int faltasPorPartido) {
+	public Player(String nombre, int edad, String equipo, double puntosPorPartido, double rebotesPorPartido,
+			double asistenciasPorPartido, double robosPorPartido, double bloqueosPorPartido, double faltasPorPartido) {
 		
 		this.nombre = nombre;
 		this.edad = edad;
@@ -50,51 +50,51 @@ public class Player {
 		this.equipo = equipo;
 	}
 
-	public int getPuntosPorPartido() {
+	public double getPuntosPorPartido() {
 		return puntosPorPartido;
 	}
 
-	public void setPuntosPorPartido(int puntosPorPartido) {
+	public void setPuntosPorPartido(double puntosPorPartido) {
 		this.puntosPorPartido = puntosPorPartido;
 	}
 
-	public int getRebotesPorPartido() {
+	public double getRebotesPorPartido() {
 		return rebotesPorPartido;
 	}
 
-	public void setRebotesPorPartido(int rebotesPorPartido) {
+	public void setRebotesPorPartido(double rebotesPorPartido) {
 		this.rebotesPorPartido = rebotesPorPartido;
 	}
 
-	public int getAsistenciasPorPartido() {
+	public double getAsistenciasPorPartido() {
 		return asistenciasPorPartido;
 	}
 
-	public void setAsistenciasPorPartido(int asistenciasPorPartido) {
+	public void setAsistenciasPorPartido(double asistenciasPorPartido) {
 		this.asistenciasPorPartido = asistenciasPorPartido;
 	}
 
-	public int getRobosPorPartido() {
+	public double getRobosPorPartido() {
 		return robosPorPartido;
 	}
 
-	public void setRobosPorPartido(int robosPorPartido) {
+	public void setRobosPorPartido(double robosPorPartido) {
 		this.robosPorPartido = robosPorPartido;
 	}
 
-	public int getBloqueosPorPartido() {
+	public double getBloqueosPorPartido() {
 		return bloqueosPorPartido;
 	}
 
-	public void setBloqueosPorPartido(int bloqueosPorPartido) {
+	public void setBloqueosPorPartido(double bloqueosPorPartido) {
 		this.bloqueosPorPartido = bloqueosPorPartido;
 	}
 
-	public int getFaltasPorPartido() {
+	public double getFaltasPorPartido() {
 		return faltasPorPartido;
 	}
 
-	public void setFaltasPorPartido(int faltasPorPartido) {
+	public void setFaltasPorPartido(double faltasPorPartido) {
 		this.faltasPorPartido = faltasPorPartido;
 	}
